odoo_pricelist_import/tests/test_pricelist_import.py
```python
from odoo.tests import common
from odoo.exceptions import UserError
import xlrd

class TestPricelistImport(common.TransactionCase):

    def setUp(self):
        super(TestPricelistImport, self).setUp()
        self.PricelistImport = self.env['pricelist.import']

    def test_01_check_xlrd_dependency(self):
        try:
            import xlrd
        except ImportError:
            self.fail("xlrd library not found")

    def test_02_import_pricelist(self):
        try:
            with open('odoo_pricelist_import/data/sample_pricelist.xlsx', 'rb') as file:
                data = file.read()
            self.PricelistImport.create({'data_file': data})
        except UserError:
            self.fail("Pricelist import failed")

    def test_03_check_imported_data(self):
        pricelist = self.env['product.pricelist'].search([('name', '=', 'Test Pricelist')])
        self.assertTrue(pricelist, "Imported pricelist not found")

    def test_04_check_imported_pricelist_items(self):
        pricelist_item = self.env['product.pricelist.item'].search([('pricelist_id.name', '=', 'Test Pricelist')])
        self.assertTrue(pricelist_item, "Imported pricelist items not found")
```